Update errors and fixed phase

+++ include/JoSIM/CCCS.hpp
@@ -32,7 +32,7 @@ class CCCS {
         const std::pair<std::string, std::string> &s,
         const std::unordered_map<std::string, int> &nm, 
         std::unordered_set<std::string> &lm,
-        std::vector<std::vector<std::pair<int, int>>> &nc,
+        std::vector<std::vector<std::pair<double, int>>> &nc,
         const std::unordered_map<JoSIM::ParameterName, Parameter> &p,
         int &branchIndex);
     void set_label(const std::string &s, std::unordered_set<std::string> &lm);
@@ -40,7 +40,7 @@ class CCCS {
       const std::pair<std::string, std::string> &n2, const std::unordered_map<std::string, int> &nm, 
       const std::string &s, int &branchIndex);
     void set_indices(const std::pair<std::string, std::string> &n1, const std::pair<std::string, std::string> &n2, 
-      const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<int, int>>> &nc, const int &branchIndex);
+      const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<double, int>>> &nc, const int &branchIndex);
     void set_currentIndex(const int &cc) { currentIndex_ = cc; }
     void set_value(const std::pair<std::string, std::string> &s, 
         const std::unordered_map<JoSIM::ParameterName, Parameter> &p);
