+++ include/JoSIM/JJ.hpp
@@ -51,7 +51,7 @@ class JJ {
         const std::pair<std::string, std::string> &s,
         const std::unordered_map<std::string, int> &nm, 
         std::unordered_set<std::string> &lm,
-        std::vector<std::vector<std::pair<int, int>>> &nc,
+        std::vector<std::vector<std::pair<double, int>>> &nc,
         const std::unordered_map<JoSIM::ParameterName, Parameter> &p,
         const std::vector<std::pair<JoSIM::Model, std::string>> &models,
         const JoSIM::AnalysisType &antyp,
@@ -61,7 +61,7 @@ class JJ {
     void set_nonZeros_and_columnIndex(const std::pair<std::string, std::string> &n, 
         const std::unordered_map<std::string, int> &nm, const std::string &s, int &branchIndex,
         const JoSIM::AnalysisType &antyp, const double &timestep);
-    void set_indices(const std::pair<std::string, std::string> &n, const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<int, int>>> &nc, const int &branchIndex);
+    void set_indices(const std::pair<std::string, std::string> &n, const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<double, int>>> &nc, const int &branchIndex);
     void set_currentIndex(const int &cc) { currentIndex_ = cc; }
     void set_variableIndex(const int &vc) { variableIndex_ = vc; }
     void set_area(const std::pair<std::string, std::string> &s,
