+++ include/JoSIM/VoltageSource.hpp
@@ -31,11 +31,11 @@ class VoltageSource {
         const std::pair<std::string, std::string> &s,
         const std::unordered_map<std::string, int> &nm, 
         std::unordered_set<std::string> &lm,
-        std::vector<std::vector<std::pair<int, int>>> &nc,
+        std::vector<std::vector<std::pair<double, int>>> &nc,
         int &branchIndex);
     void set_label(const std::string &s, std::unordered_set<std::string> &lm);
     void set_nonZeros_and_columnIndex(const std::pair<std::string, std::string> &n, const std::unordered_map<std::string, int> &nm, const std::string &s, int &branchIndex);
-    void set_indices(const std::pair<std::string, std::string> &n, const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<int, int>>> &nc, const int &branchIndex);
+    void set_indices(const std::pair<std::string, std::string> &n, const std::unordered_map<std::string, int> &nm, std::vector<std::vector<std::pair<double, int>>> &nc, const int &branchIndex);
     void set_currentIndex(const int &cc) { currentIndex_ = cc; }
     void set_sourceIndex(const int &i) { sourceIndex_ = i; }
